[#704] Fix auth redirect & private route logic
